Renamed problematic namespaces Error and *.Get
