+++ ApiEndpoints/Credentials/GetCredential.cs
@@ -6,7 +6,7 @@ namespace AnonKeyBackend.ApiEndpoints.Credentials;
 /// <summary>
 /// Handles the credentials get endpoint.
 /// </summary>
-public static class Get
+public static class GetCredential
 {
 
     /// <summary>
@@ -14,14 +14,14 @@ public static class Get
     /// </summary>
     public static Microsoft.AspNetCore.Http.HttpResults.Results<
         Ok<ApiDatastructures.Credentials.Get.CredentialsGetResponseBody>,
-        NotFound<ApiDatastructures.Error.ErrorResponseBody>,
-        BadRequest<ApiDatastructures.Error.ErrorResponseBody>>
+        NotFound<ApiDatastructures.RequestError.ErrorResponseBody>,
+        BadRequest<ApiDatastructures.RequestError.ErrorResponseBody>>
             GetGet(string credentialUuid, ClaimsPrincipal user, Data.DatabaseHandle databaseHandle)
     {
         databaseHandle.Database.EnsureCreated();
         if (user.Identity == null)
         {
-            return TypedResults.BadRequest(new ApiDatastructures.Error.ErrorResponseBody()
+            return TypedResults.BadRequest(new ApiDatastructures.RequestError.ErrorResponseBody()
             {
                 Message = "The user identity is null",
                 Detail = "The user identity is null. Did you provide a valid JWT token?"
@@ -29,7 +29,7 @@ public static class Get
         }
         if (String.IsNullOrEmpty(credentialUuid))
         {
-            return TypedResults.BadRequest(new ApiDatastructures.Error.ErrorResponseBody()
+            return TypedResults.BadRequest(new ApiDatastructures.RequestError.ErrorResponseBody()
             {
                 Message = "A credentialUuid in the request was null or an empty string",
                 Detail = "credentialUuid parameter in the request was null or an empty string. This is not allowed, please fill in all parameters."
@@ -38,7 +38,7 @@ public static class Get
 
         if (!databaseHandle.Credentials.Any(c => c.Uuid == credentialUuid))
         {
-            return TypedResults.NotFound(new ApiDatastructures.Error.ErrorResponseBody()
+            return TypedResults.NotFound(new ApiDatastructures.RequestError.ErrorResponseBody()
             {
                 Message = "No credetial with this Uuid was not found in the database",
                 Detail = "No credetial with the provided credentialUuid was not found. Please make sure the correct credentialUuid is provided."
@@ -47,7 +47,7 @@ public static class Get
 
         if (databaseHandle.Users.Single(u => u.Username == user.Identity.Name).Uuid != databaseHandle.Credentials.Single(c => c.Uuid == credentialUuid).UserUuid)
         {
-            return TypedResults.BadRequest(new ApiDatastructures.Error.ErrorResponseBody()
+            return TypedResults.BadRequest(new ApiDatastructures.RequestError.ErrorResponseBody()
             {
                 Message = "This user does not have access to this credential",
                 Detail = "The credential with the provided uuid does not belong to this user. This is not allowed, please make sure to provide a valid input."
