+++ ApiEndpoints/Folders/Create.cs
@@ -16,14 +16,14 @@ public static class Create
     /// </summary>
     public static Microsoft.AspNetCore.Http.HttpResults.Results<
         Ok<ApiDatastructures.Folders.Create.FoldersCreateResponseBody>,
-        Conflict<ApiDatastructures.Error.ErrorResponseBody>,
-        BadRequest<ApiDatastructures.Error.ErrorResponseBody>>
+        Conflict<ApiDatastructures.RequestError.ErrorResponseBody>,
+        BadRequest<ApiDatastructures.RequestError.ErrorResponseBody>>
             PostCreate(ApiDatastructures.Folders.Create.FoldersCreateRequestBody requestBody, ClaimsPrincipal user, Data.DatabaseHandle databaseHandle)
     {
         databaseHandle.Database.EnsureCreated();
         if (requestBody.Folder is null || String.IsNullOrEmpty(requestBody.Folder.Name))
         {
-            return TypedResults.BadRequest(new ApiDatastructures.Error.ErrorResponseBody()
+            return TypedResults.BadRequest(new ApiDatastructures.RequestError.ErrorResponseBody()
             {
                 Message = "The a parameter in the request was null",
                 Detail = "One of the parameters in the request was null. This is not allowed, please fill in all parameters."
@@ -32,7 +32,7 @@ public static class Create
 
         if (databaseHandle.Folders.Any(f => f.DisplayName == requestBody.Folder.Name))
         {
-            return TypedResults.Conflict(new ApiDatastructures.Error.ErrorResponseBody()
+            return TypedResults.Conflict(new ApiDatastructures.RequestError.ErrorResponseBody()
             {
                 Message = "The folder name is already taken",
                 Detail = "The folder name is already taken. Please choose a different name."
