+++ ApiEndpoints/Folders/GetFolder.cs
@@ -5,7 +5,7 @@ namespace AnonKeyBackend.ApiEndpoints.Folders;
 /// <summary>
 /// Handles the folders get endpoint. 
 /// </summary>
-public static class Get
+public static class GetFolder
 {
 
     /// <summary>
@@ -13,15 +13,15 @@ public static class Get
     /// </summary>
     public static Microsoft.AspNetCore.Http.HttpResults.Results<
         Ok<ApiDatastructures.Folders.Get.FoldersGetResponseBody>,
-        NotFound<ApiDatastructures.Error.ErrorResponseBody>,
-        BadRequest<ApiDatastructures.Error.ErrorResponseBody>>
+        NotFound<ApiDatastructures.RequestError.ErrorResponseBody>,
+        BadRequest<ApiDatastructures.RequestError.ErrorResponseBody>>
             GetGet(string folderUuid, ClaimsPrincipal user, Data.DatabaseHandle databaseHandle)
     {
         databaseHandle.Database.EnsureCreated();
 
         if (user.Identity == null)
         {
-            return TypedResults.BadRequest(new ApiDatastructures.Error.ErrorResponseBody()
+            return TypedResults.BadRequest(new ApiDatastructures.RequestError.ErrorResponseBody()
             {
                 Message = "The user identity is null",
                 Detail = "The user identity is null. Did you provide a valid JWT token?"
@@ -29,7 +29,7 @@ public static class Get
         }
         if (String.IsNullOrEmpty(folderUuid))
         {
-            return TypedResults.BadRequest(new ApiDatastructures.Error.ErrorResponseBody()
+            return TypedResults.BadRequest(new ApiDatastructures.RequestError.ErrorResponseBody()
             {
                 Message = "The folder UUID is missing",
                 Detail = "The folder UUID is missing from the request. Please provide the folder UUID."
@@ -40,7 +40,7 @@ public static class Get
         Folder? folder = databaseHandle.Folders.FirstOrDefault(f => f.Uuid == folderUuid);
         if (folder == null || userObject == null || folder.UserUuid != userObject.Uuid)
         {
-            return TypedResults.NotFound(new ApiDatastructures.Error.ErrorResponseBody()
+            return TypedResults.NotFound(new ApiDatastructures.RequestError.ErrorResponseBody()
             {
                 Message = "The folder does not exist",
                 Detail = "The folder with the given UUID does not exist."
