+++ ApiEndpoints/Users/EndpointSetup.cs
@@ -12,7 +12,7 @@ public static class EndpointSetup
     public static void MapEndpoints(WebApplication app)
     {
         app.MapPost("/user/create", Create.PostCreate).WithTags("Users").WithOpenApi();
-        app.MapGet("/user/get", Get.GetGet).WithTags("Users").WithOpenApi().RequireAuthorization("user");
+        app.MapGet("/user/get", GetUser.GetGet).WithTags("Users").WithOpenApi().RequireAuthorization("user");
         app.MapPut("/user/update", Update.PutUpdate).WithTags("Users").WithOpenApi().RequireAuthorization("user");
         app.MapDelete("/user/delete", Delete.DeleteDelete).WithTags("Users").WithOpenApi().RequireAuthorization("user");
     }
